requirements.txt
agno
openai
python-dotenv
requests
openpyxl